Upgraded coinpaprika to v3 framework
