+++ .changeset/hip-penguins-warn.md
@@ -1,0 +1,5 @@
+---
+'@chainlink/coinpaprika-adapter': major
+---
+
+Upgraded adapter from v2 to v3 framework
