+++ packages/core/legos/src/sources.ts
@@ -38,7 +38,6 @@ import * as coinlore from '@chainlink/coinlore-adapter'
 import * as coinmarketcap from '@chainlink/coinmarketcap-adapter'
 import * as coinmetrics from '@chainlink/coinmetrics-adapter'
 import * as coinpaprika from '@chainlink/coinpaprika-adapter'
-import * as coinpaprika_test from '@chainlink/coinpaprika-test-adapter'
 import * as coinranking from '@chainlink/coinranking-adapter'
 import * as covid_tracker from '@chainlink/covid-tracker-adapter'
 import * as cryptex from '@chainlink/cryptex-adapter'
@@ -201,7 +200,6 @@ export default {
   coinmarketcap,
   coinmetrics,
   coinpaprika,
-  coinpaprika_test,
   coinranking,
   covid_tracker,
   cryptex,
