+++ packages/core/legos/tsconfig.json
@@ -131,9 +131,6 @@
       "path": "../../sources/coinpaprika"
     },
     {
-      "path": "../../sources/coinpaprika-test"
-    },
-    {
       "path": "../../sources/coinranking"
     },
     {
