--- packages/sources/coinpaprika-test/CHANGELOG.md
@@ -1,94 +1,0 @@
-# @chainlink/coinpaprika-test-adapter
-
-## 1.2.4
-
-### Patch Changes
-
-- bfa201d6d: Bumped framework version
-- bfa201d6d: Bumped framework version
-- c600ca386: Upgrade typescript version to 5.0.4
-- bfa201d6d: Bumped framework version
-
-## 1.2.3
-
-### Patch Changes
-
-- 9bcb13c90: Bumped framework version
-
-## 1.2.2
-
-### Patch Changes
-
-- 31af84e69: Bumped framework version
-
-## 1.2.1
-
-### Patch Changes
-
-- 77ad946a9: Bumped framework version
-
-## 1.2.0
-
-### Minor Changes
-
-- dfda2a99c: Refactor to unbatch quotes and consolidate endpoints
-
-## 1.1.6
-
-### Patch Changes
-
-- 4245b4d8b: Bumped framework version
-
-## 1.1.5
-
-### Patch Changes
-
-- 8fe5be225: Bumped framework version
-
-## 1.1.4
-
-### Patch Changes
-
-- c9cee149e: Bump fw version
-
-## 1.1.3
-
-### Patch Changes
-
-- 2fdaa5aa4: Bump v3 version
-
-## 1.1.2
-
-### Patch Changes
-
-- bc8886c63: Bumped framework version
-
-## 1.1.1
-
-### Patch Changes
-
-- 37c0d8883: Bumped v3 framework version
-
-## 1.1.0
-
-### Minor Changes
-
-- d474a5553: Update v3 framework to latest version
-
-## 1.0.2
-
-### Patch Changes
-
-- 65014014d: Bumped v3 framework version. Upgraded typescript to version 4.9.5
-
-## 1.0.1
-
-### Patch Changes
-
-- c152ed06e: Bumped v3 framework version
-
-## 1.0.0
-
-### Major Changes
-
-- 23a1bd4b3: Added coinpaprika-test EA
