--- packages/sources/coinpaprika-test/src/config/index.ts
@@ -1,29 +1,0 @@
-import { AdapterConfig } from '@chainlink/external-adapter-framework/config'
-
-export const DEFAULT_API_ENDPOINT = 'https://api.coinpaprika.com'
-export const PRO_API_ENDPOINT = 'https://api-pro.coinpaprika.com'
-
-export const config = new AdapterConfig({
-  API_ENDPOINT: {
-    description: 'The HTTP URL to retrieve data from',
-    type: 'string',
-    required: false,
-  },
-  API_KEY: {
-    description: 'An API key for Coinpaprika',
-    type: 'string',
-    required: false,
-    sensitive: true,
-  },
-})
-
-export const getApiEndpoint = (settings: typeof config.settings): string =>
-  settings.API_ENDPOINT || (settings.API_KEY ? PRO_API_ENDPOINT : DEFAULT_API_ENDPOINT)
-
-export const getApiHeaders = (settings: typeof config.settings): { Authorization?: string } => {
-  const headers: { Authorization?: string } = {}
-  if (settings.API_KEY) {
-    headers['Authorization'] = settings.API_KEY
-  }
-  return headers
-}
