--- packages/sources/coinpaprika-test/src/endpoint/index.ts
@@ -1,4 +1,0 @@
-export { endpoint as coins } from './coins'
-export { endpoint as crypto } from './crypto'
-export { endpoint as global } from './global'
-export { endpoint as vwap } from './vwap'
