--- packages/sources/coinpaprika-test/test/integration/adapter.test.ts
@@ -1,175 +1,0 @@
-import { mockCryptoResponseSuccess } from './fixtures'
-import { SuperTest, Test } from 'supertest'
-import { setupExternalAdapterTest, SuiteContext } from './setup'
-import { ServerInstance } from '@chainlink/external-adapter-framework'
-
-describe('execute', () => {
-  let spy: jest.SpyInstance
-  beforeAll(async () => {
-    const mockDate = new Date('2022-01-01T11:11:11.111Z')
-    spy = jest.spyOn(Date, 'now').mockReturnValue(mockDate.getTime())
-  })
-
-  afterAll((done) => {
-    spy.mockRestore()
-    done()
-  })
-  const id = '1'
-
-  const context: SuiteContext = {
-    req: null,
-    server: async () => {
-      // workaround for failing integration tests that run in parallel
-      process.env['RATE_LIMIT_CAPACITY_SECOND'] = '10000'
-      process.env['METRICS_ENABLED'] = 'false'
-      const server = (await import('../../src')).server
-      return server() as Promise<ServerInstance>
-    },
-  }
-
-  const envVariables = {
-    CACHE_ENABLED: 'false',
-  }
-
-  setupExternalAdapterTest(envVariables, context)
-
-  describe('crypto batch endpoint', () => {
-    const data = {
-      id,
-      data: {
-        base: 'ETH',
-        quote: 'USD',
-      },
-    }
-
-    it('should return success', async () => {
-      mockCryptoResponseSuccess()
-
-      const response = await (context.req as SuperTest<Test>)
-        .post('/')
-        .send(data)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(200)
-      expect(response.body).toMatchSnapshot()
-    })
-  })
-
-  describe('crypto volume endpoint', () => {
-    const data = {
-      id,
-      data: {
-        base: 'ETH',
-        quote: 'USD',
-        endpoint: 'volume',
-      },
-    }
-
-    it('should return success', async () => {
-      mockCryptoResponseSuccess()
-
-      const response = await (context.req as SuperTest<Test>)
-        .post('/')
-        .send(data)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(200)
-      expect(response.body).toMatchSnapshot()
-    })
-  })
-
-  describe('crypto marketcap endpoint', () => {
-    const data = {
-      id,
-      data: {
-        base: 'ETH',
-        quote: 'USD',
-        endpoint: 'marketcap',
-      },
-    }
-
-    it('should return success', async () => {
-      mockCryptoResponseSuccess()
-
-      const response = await (context.req as SuperTest<Test>)
-        .post('/')
-        .send(data)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(200)
-      expect(response.body).toMatchSnapshot()
-    })
-  })
-
-  describe('dominance endpoint', () => {
-    const data = {
-      id,
-      data: {
-        market: 'BTC',
-        endpoint: 'dominance',
-      },
-    }
-
-    it('should return success', async () => {
-      mockCryptoResponseSuccess()
-
-      const response = await (context.req as SuperTest<Test>)
-        .post('/')
-        .send(data)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(200)
-      expect(response.body).toMatchSnapshot()
-    })
-  })
-
-  describe('globalmarketcap endpoint', () => {
-    const data = {
-      id,
-      data: {
-        market: 'USD',
-        endpoint: 'globalmarketcap',
-      },
-    }
-
-    it('should return success', async () => {
-      mockCryptoResponseSuccess()
-
-      const response = await (context.req as SuperTest<Test>)
-        .post('/')
-        .send(data)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(200)
-      expect(response.body).toMatchSnapshot()
-    })
-  })
-
-  describe('vwap endpoint', () => {
-    const data = {
-      id,
-      data: {
-        base: 'ETH',
-        endpoint: 'vwap',
-      },
-    }
-
-    it('should return success', async () => {
-      mockCryptoResponseSuccess()
-
-      const response = await (context.req as SuperTest<Test>)
-        .post('/')
-        .send(data)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(200)
-      expect(response.body).toMatchSnapshot()
-    })
-  })
-})
