--- packages/sources/coinpaprika/schemas/env.json
@@ -1,20 +1,0 @@
-{
-  "$id": "https://external-adapters.chainlinklabs.com/schemas/coinpaprika-adapter.json",
-  "title": "Chainlink External Adapter for CoinPaprika",
-  "description": "_Note: the `-single` endpoints have the same functionality as their original endpoint, except they will only fetch data for the single asset being queried._\n\n If network spikes are observed with the coinpaprika adapter, it may be due to not providing a coin-id and instead using a ticker symbol. Ticker symbols must be converted to coin-ids before they are fetched from the data provider. This conversion step requires downloading a large list of every single coin-id and matching ticker symbol from the data provider, impacting network performance. Provide a coin-id directly using the coin-id parameter, or provide a coin-id via the overrides parameter.\n If the adapter still has large spikes in network usage after using the coin-id instead of a ticker symbol, this because, in order to handle requests for multiple price pairs, the data provider requires fetching every price pair the support simultaneously. The resulting response payload is very large (~2 MB), leading to impacted network performance.\n  A solution is to use the `crypto-single` endpoint instead to only fetch single price pairs as opposed to all price pairs. This will work, however batching will not be supported.\n  This is because, by using the crypto-single endpoint, an individual request is sent to the data provider to fetch each price pair, as opposed to one request for all price pairs. Thus, the total number of API requests is increased by a factor of the number of feeds, but each request will have a much smaller payload.",
-  "required": [],
-  "type": "object",
-  "properties": {
-    "API_KEY": {
-      "required": false
-    },
-    "IS_TEST_MODE": {
-      "required": false
-    }
-  },
-  "allOf": [
-    {
-      "$ref": "https://external-adapters.chainlinklabs.com/schemas/ea-bootstrap.json"
-    }
-  ]
-}
