--- packages/sources/coinpaprika/src/config/limits.json
@@ -1,24 +1,0 @@
-{
-  "http": {
-    "free": {
-      "rateLimit1s": 10,
-      "rateLimit1h": 69.44,
-      "note": "50k/mo for free"
-    },
-    "pro": {
-      "rateLimit1s": 10,
-      "rateLimit1h": 347.22,
-      "note": "250k/mo for pro"
-    },
-    "business": {
-      "rateLimit1s": 10,
-      "rateLimit1h": 1388.888,
-      "note": "1mil/mo for business"
-    },
-    "enterprise": {
-      "rateLimit1s": 10,
-      "note": "unlimited per month"
-    }
-  },
-  "ws": {}
-}
