+++ packages/sources/coinpaprika/src/endpoint/global.ts
