+++ packages/sources/coinpaprika/src/endpoint/index.ts
@@ -1,21 +1,4 @@
-import { TInputParameters as CryptoInputParameters } from './crypto'
-import { TInputParameters as CryptoSingleInputParameters } from './crypto-single'
-import { TInputParameters as DominanceInputParameters } from './dominance'
-import { TInputParameters as GlobalMcapInputParameters } from './globalMarketcap'
-import { TInputParameters as CoinsInputParameters } from './coins'
-import { TInputParameters as VwapInputParameters } from './vwap'
-
-export type TInputParameters =
-  | CryptoInputParameters
-  | CryptoSingleInputParameters
-  | DominanceInputParameters
-  | GlobalMcapInputParameters
-  | CoinsInputParameters
-  | VwapInputParameters
-
-export * as crypto from './crypto'
-export * as cryptoSingle from './crypto-single'
-export * as dominance from './dominance'
-export * as globalMarketcap from './globalMarketcap'
-export * as coins from './coins'
-export * as vwap from './vwap'
+export { endpoint as coins } from './coins'
+export { endpoint as crypto } from './crypto'
+export { endpoint as global } from './global'
+export { endpoint as vwap } from './vwap'
