+++ packages/sources/coinpaprika/src/global-utils.ts
