+++ packages/sources/coinpaprika/src/index.ts
@@ -1,10 +1,36 @@
-import { expose } from '@chainlink/ea-bootstrap'
-import { endpointSelector, makeExecute } from './adapter'
-import * as endpoints from './endpoint'
-import { makeConfig, NAME } from './config'
-import rateLimit from './config/limits.json'
+import { expose, ServerInstance } from '@chainlink/external-adapter-framework'
+import { PriceAdapter } from '@chainlink/external-adapter-framework/adapter'
+import { config } from './config'
+import { coins, crypto, global, vwap } from './endpoint'
 
-const adapterContext = { name: NAME, rateLimit }
+export const adapter = new PriceAdapter({
+  defaultEndpoint: crypto.name,
+  name: 'COINPAPRIKA',
+  config,
+  rateLimiting: {
+    tiers: {
+      free: {
+        rateLimit1s: 10,
+        rateLimit1h: 69.44,
+        note: '50k/mo for free',
+      },
+      pro: {
+        rateLimit1s: 10,
+        rateLimit1h: 347.22,
+        note: '250k/mo for pro',
+      },
+      business: {
+        rateLimit1s: 10,
+        rateLimit1h: 1388.888,
+        note: '1mil/mo for business',
+      },
+      enterprise: {
+        rateLimit1s: 10,
+        note: 'unlimited per month',
+      },
+    },
+  },
+  endpoints: [crypto, global, coins, vwap],
+})
 
-const { server } = expose(adapterContext, makeExecute(), undefined, endpointSelector)
-export { NAME, makeExecute, makeConfig, server, endpoints }
+export const server = (): Promise<ServerInstance | undefined> => expose(adapter)
