+++ packages/sources/coinpaprika/test-payload.json
@@ -1,6 +1,6 @@
 {
  "requests": [{
-  "from": "SUSHI",
-  "to": "ETH"
+  "from": "LINK",
+  "to": "USD"
  }]
 }
