+++ packages/sources/coinpaprika/test/integration/fixtures.ts
@@ -4,6 +4,7 @@ export const mockCryptoResponseSuccess = (): nock.Scope =>
   nock('https://api.coinpaprika.com', {
     encodedQueryParams: true,
   })
+    .persist()
     .get('/v1/tickers')
     .query({ quotes: 'USD' })
     .reply(
@@ -22,7 +23,7 @@ export const mockCryptoResponseSuccess = (): nock.Scope =>
           last_updated: '2021-10-22T18:11:05Z',
           quotes: {
             USD: {
-              price: 3949.2425813062,
+              price: 4949.2425813062,
               volume_24h: 24136641726.138,
               volume_24h_change_24h: -35.07,
               market_cap: 466143026900,
@@ -120,6 +121,7 @@ export const mockCryptoResponseSuccess = (): nock.Scope =>
         'Origin',
       ],
     )
+    .persist()
     .get(`/v1/global`)
     .reply(
       200,
@@ -149,6 +151,7 @@ export const mockCryptoResponseSuccess = (): nock.Scope =>
         'Origin',
       ],
     )
+    .persist()
     .get('/v1/coins')
     .reply(200, [
       {
@@ -188,6 +191,7 @@ export const mockCryptoResponseSuccess = (): nock.Scope =>
         type: 'token',
       },
     ])
+    .persist()
     .get('/v1/tickers/eth-ethereum')
     .query({ quotes: 'USD' })
     .reply(
@@ -236,6 +240,7 @@ export const mockCryptoResponseSuccess = (): nock.Scope =>
         'Origin',
       ],
     )
+    .persist()
     .get('/v1/tickers/fake-btc-bitcoin')
     .query({ quotes: 'USD' })
     .reply(
@@ -284,6 +289,7 @@ export const mockCryptoResponseSuccess = (): nock.Scope =>
         'Origin',
       ],
     )
+    .persist()
     .get('/v1/tickers/ampl-ampleforth/historical')
     .query({
       start: /^\d{4}-\d{2}-\d{2}$/,
@@ -297,6 +303,7 @@ export const mockCryptoResponseSuccess = (): nock.Scope =>
         market_cap: 106686649,
       },
     ])
+    .persist()
     .get('/v1/tickers/eth-ethereum/historical')
     .query({
       start: /^\d{4}-\d{2}-\d{2}$/,
@@ -310,264 +317,4 @@ export const mockCryptoResponseSuccess = (): nock.Scope =>
         market_cap: 999999999,
       },
     ])
-
-export const mockPROCryptoResponseSuccess = (): nock.Scope =>
-  nock('https://api-pro.coinpaprika.com', {
-    encodedQueryParams: true,
-    reqheaders: {
-      Authorization: 'fake-api-key',
-    },
-  })
-    .get('/v1/tickers')
-    .query({ quotes: 'USD' })
-    .reply(
-      200,
-      () => [
-        {
-          id: 'eth-ethereum',
-          name: 'Ethereum',
-          symbol: 'ETH',
-          rank: 2,
-          circulating_supply: 118033526,
-          total_supply: 118033574,
-          max_supply: 0,
-          beta_value: 1.08967,
-          first_data_at: '2015-08-07T00:00:00Z',
-          last_updated: '2021-10-22T18:11:05Z',
-          quotes: {
-            USD: {
-              price: 3949.2425813062,
-              volume_24h: 24136641726.138,
-              volume_24h_change_24h: -35.07,
-              market_cap: 466143026900,
-              market_cap_change_24h: -3.44,
-              percent_change_15m: 0.14,
-              percent_change_30m: -0.18,
-              percent_change_1h: -0.64,
-              percent_change_6h: -4.09,
-              percent_change_12h: -4.65,
-              percent_change_24h: -3.45,
-              percent_change_7d: 2.23,
-              percent_change_30d: 28.11,
-              percent_change_1y: 844.08,
-              ath_price: 4365.2053035,
-              ath_date: '2021-05-12T06:06:20Z',
-              percent_from_price_ath: -9.53,
-            },
-          },
-        },
-      ],
-      [
-        'Content-Type',
-        'application/json',
-        'Connection',
-        'close',
-        'Vary',
-        'Accept-Encoding',
-        'Vary',
-        'Origin',
-      ],
-    )
-
-export const mockCryptoSingleResponseSuccess = (): nock.Scope =>
-  nock('https://api.coinpaprika.com', {
-    encodedQueryParams: true,
-  })
-    .get('/v1/coins')
-    .reply(200, [
-      {
-        id: 'aaaa-token',
-        name: 'AAAA token',
-        symbol: 'AAAA',
-        rank: 200,
-        is_new: false,
-        is_active: true,
-        type: 'token',
-      },
-      {
-        id: 'ampl-ampleforth',
-        name: 'Ampleforth',
-        symbol: 'AMPL',
-        rank: 289,
-        is_new: false,
-        is_active: true,
-        type: 'token',
-      },
-      {
-        id: 'fake-ampleforth',
-        name: 'Fakeampleforth',
-        symbol: 'AMPL',
-        rank: 300,
-        is_new: false,
-        is_active: true,
-        type: 'token',
-      },
-      {
-        id: 'real-bbbb',
-        name: 'Realbbbb',
-        symbol: 'BBBB',
-        rank: 300,
-        is_new: false,
-        is_active: true,
-        type: 'token',
-      },
-      {
-        id: 'fake-bbbb',
-        name: 'Fakebbbb',
-        symbol: 'BBBB',
-        rank: 300,
-        is_new: false,
-        is_active: true,
-        type: 'token',
-      },
-    ])
-    .get('/v1/tickers/eth-ethereum')
-    .query({ quotes: 'USD' })
-    .reply(
-      200,
-      () => [
-        {
-          id: 'eth-ethereum',
-          name: 'Ethereum',
-          symbol: 'ETH',
-          rank: 2,
-          circulating_supply: 118033526,
-          total_supply: 118033574,
-          max_supply: 0,
-          beta_value: 1.08967,
-          first_data_at: '2015-08-07T00:00:00Z',
-          last_updated: '2021-10-22T18:11:05Z',
-          quotes: {
-            USD: {
-              price: 3949.2425813062,
-              volume_24h: 24136641726.138,
-              volume_24h_change_24h: -35.07,
-              market_cap: 466143026900,
-              market_cap_change_24h: -3.44,
-              percent_change_15m: 0.14,
-              percent_change_30m: -0.18,
-              percent_change_1h: -0.64,
-              percent_change_6h: -4.09,
-              percent_change_12h: -4.65,
-              percent_change_24h: -3.45,
-              percent_change_7d: 2.23,
-              percent_change_30d: 28.11,
-              percent_change_1y: 844.08,
-              ath_price: 4365.2053035,
-              ath_date: '2021-05-12T06:06:20Z',
-              percent_from_price_ath: -9.53,
-            },
-          },
-        },
-      ],
-      [
-        'Content-Type',
-        'application/json',
-        'Connection',
-        'close',
-        'Vary',
-        'Accept-Encoding',
-        'Vary',
-        'Origin',
-      ],
-    )
-    .get('/v1/tickers/aaaa-token')
-    .query({ quotes: 'USD' })
-    .reply(
-      200,
-      () => [
-        {
-          id: 'aaaa-token',
-          name: 'AAAA token',
-          symbol: 'AAAA',
-          rank: 2,
-          circulating_supply: 118033526,
-          total_supply: 118033574,
-          max_supply: 0,
-          beta_value: 1.08967,
-          first_data_at: '2015-08-07T00:00:00Z',
-          last_updated: '2021-10-22T18:11:05Z',
-          quotes: {
-            USD: {
-              price: 200.0,
-              volume_24h: 24136641726.138,
-              volume_24h_change_24h: -35.07,
-              market_cap: 466143026900,
-              market_cap_change_24h: -3.44,
-              percent_change_15m: 0.14,
-              percent_change_30m: -0.18,
-              percent_change_1h: -0.64,
-              percent_change_6h: -4.09,
-              percent_change_12h: -4.65,
-              percent_change_24h: -3.45,
-              percent_change_7d: 2.23,
-              percent_change_30d: 28.11,
-              percent_change_1y: 844.08,
-              ath_price: 4365.2053035,
-              ath_date: '2021-05-12T06:06:20Z',
-              percent_from_price_ath: -9.53,
-            },
-          },
-        },
-      ],
-      [
-        'Content-Type',
-        'application/json',
-        'Connection',
-        'close',
-        'Vary',
-        'Accept-Encoding',
-        'Vary',
-        'Origin',
-      ],
-    )
-    .get('/v1/tickers/real-bbbb')
-    .query({ quotes: 'USD' })
-    .reply(
-      200,
-      () => [
-        {
-          id: 'real-bbbb',
-          name: 'Realbbbb',
-          symbol: 'BBBB',
-          rank: 2,
-          circulating_supply: 118033526,
-          total_supply: 118033574,
-          max_supply: 0,
-          beta_value: 1.08967,
-          first_data_at: '2015-08-07T00:00:00Z',
-          last_updated: '2021-10-22T18:11:05Z',
-          quotes: {
-            USD: {
-              price: 300.0,
-              volume_24h: 24136641726.138,
-              volume_24h_change_24h: -35.07,
-              market_cap: 466143026900,
-              market_cap_change_24h: -3.44,
-              percent_change_15m: 0.14,
-              percent_change_30m: -0.18,
-              percent_change_1h: -0.64,
-              percent_change_6h: -4.09,
-              percent_change_12h: -4.65,
-              percent_change_24h: -3.45,
-              percent_change_7d: 2.23,
-              percent_change_30d: 28.11,
-              percent_change_1y: 844.08,
-              ath_price: 4365.2053035,
-              ath_date: '2021-05-12T06:06:20Z',
-              percent_from_price_ath: -9.53,
-            },
-          },
-        },
-      ],
-      [
-        'Content-Type',
-        'application/json',
-        'Connection',
-        'close',
-        'Vary',
-        'Accept-Encoding',
-        'Vary',
-        'Origin',
-      ],
-    )
+    .persist()
