+++ packages/tsconfig.json
@@ -243,9 +243,6 @@
       "path": "./sources/coinpaprika"
     },
     {
-      "path": "./sources/coinpaprika-test"
-    },
-    {
       "path": "./sources/coinranking"
     },
     {
