+++ packages/tsconfig.test.json
@@ -243,9 +243,6 @@
       "path": "./sources/coinpaprika/tsconfig.test.json"
     },
     {
-      "path": "./sources/coinpaprika-test/tsconfig.test.json"
-    },
-    {
       "path": "./sources/coinranking/tsconfig.test.json"
     },
     {
